[feat] code injection 방지

+++ app/controller/mainController.php
@@ -85,7 +85,7 @@ class MainController
      */
     public static function postSignup()
     {
-        $data = json_decode(file_get_contents("php://input"));
+        $data = json_decode(strip_tags(file_get_contents("php://input")));
 
         $email = $data->email;
         $username = $data->username;
@@ -132,12 +132,13 @@ class MainController
      */
     public static function postSignin()
     {
-        $data = json_decode(file_get_contents("php://input"));
+        $data = json_decode(strip_tags(file_get_contents("php://input")));
 
         $email = $data->email;
         $password = $data->password;
 
         $response = self::postSigninProcess($email, $password);
+        $response['zzz'] = strip_tags(file_get_contents("php://input"));
         return print_r(json_encode($response));
     }
 
@@ -209,6 +210,12 @@ class MainController
         $username = $data->username;
         $baseImage = $data->baseImage;
         $stickyImages = $data->stickyImages;
+
+        if ($username !== $_SESSION['username'])
+        {
+            http_response_code(401);
+            return;
+        }
         
         //파일 만들기
         $userId = self::getUserIdbyUsername($username);
@@ -257,10 +264,13 @@ class MainController
         $data = json_decode(file_get_contents("php://input"));
 
         $imageId = str_replace("captured-image-", "", $data->imageId);
+        $username = $data->username;
         $result = self::getModel()->postImage($imageId);
         $body = array();
 
-        if ($result['message'] === '중복')
+        if ($username !== $_SESSION['username'])
+            http_response_code(401);
+        else if ($result['message'] === '중복')
         {
             http_response_code(409);
             $body['message'] = '이미 업로드한 이미지입니다.';
@@ -296,11 +306,11 @@ class MainController
         $username = $data->username;
 
         if (!isset($_SESSION['username']))
-        {
             http_response_code(400);
-            return;
-        }
-        self::postLikesProcess($postId, $username);
+        else if ($username !== $_SESSION['username'])
+            http_response_code(401);
+        else
+            self::postLikesProcess($postId, $username);
         return;
     }
     public static function postLikesProcess($postId, $username)
@@ -323,10 +333,15 @@ class MainController
     {
         $data = json_decode(file_get_contents("php://input"));
 
-        $comment = $data->comment;
+        $comment = htmlspecialchars($data->comment);
         $postId = $data->postId;
         $username = $data->username;
 
+        if ($username !== $_SESSION['username'])
+        {
+            http_response_code(401);
+            return;
+        }
         self::getModel()->postComment($comment, $postId, $username);
         http_response_code(201);
         return;
@@ -354,9 +369,15 @@ class MainController
      */
     public static function patchUsername()
     {
-        $data = json_decode(file_get_contents("php://input"));
-        $change = $data->username;
-        $username = $_SESSION['username'];
+        $data = json_decode(strip_tags(file_get_contents("php://input")));
+        $change = $data->change;
+        $username = $data->username;
+        
+        if ($username !== $_SESSION['username'])
+        {
+            http_response_code(401);
+            return;
+        }
         
         $body = self::patchUsernameProcess($change, $username);
         return print_r(json_encode($body));
@@ -388,9 +409,16 @@ class MainController
      */
     public static function patchEmail()
     {
-        $data = json_decode(file_get_contents("php://input"));
+        $data = json_decode(strip_tags(file_get_contents("php://input")));
         $change = $data->email;
         $email = $_SESSION['email'];
+        $username = $data->username;
+        
+        if ($username !== $_SESSION['username'])
+        {
+            http_response_code(401);
+            return;
+        }
 
         $body = self::patchEmailProcess($change, $email);
         return print_r(json_encode($body));
@@ -421,11 +449,17 @@ class MainController
      */
     public static function patchPassword()
     {
-        $data = json_decode(file_get_contents("php://input"));
+        $data = json_decode(strip_tags(file_get_contents("php://input")));
         $originPassword = $data->originPassword;
         $newPassword = $data->newPassword;
         $checkPassword = $data->checkPassword;
-
+        $username = $data->username;
+        
+        if ($username !== $_SESSION['username'])
+        {
+            http_response_code(401);
+            return;
+        }
         $body = self::patchPasswordProcess($originPassword, $newPassword, $checkPassword);
         return print_r(json_encode($body));
     }
@@ -483,6 +517,13 @@ class MainController
         $data = json_decode(file_get_contents("php://input"));
 
         $commentId = $data->commentId;
+        $username = $data->username;
+
+        if ($username !== $_SESSION['username'])
+        {
+            http_response_code(401);
+            return;
+        }
         
         $result = self::getModel()->deleteComment($commentId);
         http_response_code(200);
@@ -497,8 +538,14 @@ class MainController
         $data = json_decode(file_get_contents("php://input"));
 
         $commentId = $data->commentId;
-        $newComment = $data->newComment;
+        $newComment = htmlspecialchars($data->newComment);
+        $username = $data->username;
 
+        if ($username !== $_SESSION['username'])
+        {
+            http_response_code(401);
+            return;
+        }
         $result = self::getModel()->patchComment($commentId, $newComment);
         http_response_code(200);
         return;
@@ -520,6 +567,13 @@ class MainController
         $data = json_decode(file_get_contents("php://input"));
 
         $postId = $data->postId;
+        $username = $data->username;
+
+        if ($username !== $_SESSION['username'])
+        {
+            http_response_code(401);
+            return;
+        }
 
         $result = self::getModel()->deletePost($postId);
         http_response_code(200);
@@ -534,6 +588,13 @@ class MainController
         $data = json_decode(file_get_contents("php://input"));
 
         $imageId = $data->imageId;
+        $username = $data->username;
+
+        if ($username !== $_SESSION['username'])
+        {
+            http_response_code(401);
+            return;
+        }
         $image = self::getModel()->getImageByImageId($imageId)['data'];
         unlink($image);
